+++ src/frontend/app/components/Map/helpers.js
@@ -481,90 +481,6 @@ export function getDRA(coords, point, map) {
     });
 }
 
-const MajorPopulationCenterTypes = [
-  // 'Locality', // < 50
-  // 'Community', // > 50, unincorporated
-  'Village', // < 2,500
-  'Village (1)',
-  'Village (2)',
-  'First Nation Village',
-  'Town', // < 5,000
-  'District Municipality (1)', // > 800 hectares, < 5 people/hectare
-  'City', // > 5,000
-];
-
-const MinorPopulationCenterTypes = [
-  // 'Locality', // < 50
-  'Community', // > 50, unincorporated
-];
-
-async function filterByTypes(features, types, fromCoords, reverseRouteOrder) {
-  const results = [];
-  for (const feature of features) {
-    const index = types.indexOf(feature.properties.featureType);
-    if (index < 0) { continue; }
-
-    const toCoords = feature.geometry.coordinates;
-    let direction = getCardinalDirection(fromCoords, toCoords, true);
-    if (reverseRouteOrder) {
-      direction = getCardinalDirection(toCoords, fromCoords, true);
-    }
-    const route = await getNonDirectionalRoute(fromCoords, toCoords);
-
-    if (!route || route.distance < 0) { continue; }
-
-    results.push({
-      id: `bcgnws-${feature.properties.feature.id}`,
-      source: "bcgnws",
-      name: feature.properties.name,
-      type: feature.properties.featureType,
-      coordinates: feature.geometry.coordinates,
-      coords: feature.geometry.coordinates,
-      distance: route.distance,
-      direction,
-      phrase: `${Math.round(route.distance)}km ${direction} of ${feature.properties.name}`,
-      size: types === MajorPopulationCenterTypes ? 'major' : 'minor',
-    });
-  }
-  return results;
-};
-
-export async function getNearbyFromBCGNWS(coords, reverseRouteOrder) {
-  try {
-    const baseUrl = "https://apps.gov.bc.ca/pub/bcgnws/names/near";
-    const params = {
-      featureClass: 1,
-      official: 1,
-      itemsPerPage: 200,
-      startIndex: 1,
-      featurePoint: coords[0] + "," + coords[1],
-      distance: 100, // km
-      outputFormat: "json",
-      outputStyle: "detail",
-      outputSRS: 4326
-    };
-    const url = new URL(baseUrl);
-    url.search = new URLSearchParams(params).toString();
-    const apiUrl = url.toString();
-
-    const response = await fetch(apiUrl, {mode: 'cors'});
-    if (!response.ok) {
-      const text = await response.text();
-      const message = (text.match(/<p><b>Message<\/b>(.*?)<\/p>/imv) || ['', 'No detail provided'])[1];
-      throw Error(`BCGNWS request failed (status ${response.status})`, { cause: message.trim() });
-    }
-    const data = await response.json();
-    let results = await filterByTypes(data.features, MajorPopulationCenterTypes, coords, reverseRouteOrder);
-    if (results.length === 0) {
-      results.push(...await filterByTypes(data.features, MinorPopulationCenterTypes, coords, reverseRouteOrder));
-    }
-    results.sort((a, b) => a.distance - b.distance);
-    return results.map((result) => ({... result, source: 'bcgnws', displayDistance: Math.round(result.distance)})).slice(0, 6);
-  } catch (err) {
-    console.error(err);
-    return [{ source: 'bcgnws', name: 'error', phrase: 'Problem retrieving population centres data.' }];
-  }
-}
 
 export const convertToDateTimeLocalString = (date) => {
   if (!date) return '';
