+++ src/frontend/app/events/forms/forms.scss
@@ -208,7 +208,9 @@
     }
 
     .handle {
-      flex: 0; cursor: pointer; user-select: none;
+      flex: 0;
+      cursor: grab;
+      user-select: none;
       font-size: 1rem;
       font-weight: bold;
       padding: 0 0.25rem;
@@ -331,6 +333,11 @@
             }
 
             &.landmarks {
+              .highlight {
+                font-weight: bold;
+                color: $BC-Blue;
+              }
+
               .control {
                 flex: 1;
                 display: flex;
@@ -344,12 +351,17 @@
                 display: flex;
                 justify-content: space-between;
                 align-items: first baseline;
-                cursor: grab;
+                cursor: default;
                 background: white;
 
                 margin-bottom: 0.5rem;
 
+                &.no-results {
+                  font-style: italic;
+                }
+
                 .icon {
+                  cursor: pointer;
                   img {
                     margin-bottom: -0.1rem;
                   }
@@ -382,9 +394,10 @@
                 .phrase {
                   flex: 1;
                   text-align: left;
+                  white-space: pre-wrap;
 
-                  display: flex;
                   input { flex: 1; }
+                  &.phrase-other { display: flex; }
                 }
 
                 .icon {
@@ -431,18 +444,25 @@
 
               .search-box {
                 position: relative;
+
                 .search-icon {
                   position: absolute;
                   right: 10px;
                   top: 0.2rem;
                   display: inline-block;
-                  color: #ccc;
+                    color: #ccc;
 
                   svg {
                     height: 14px;
                     width: 14px;
                   }
                 }
+                input:focus + button.search-icon {
+                  svg {
+                    color: black !important;
+                  }
+                }
+
               }
               .landmark-search {
                 width: 100%;
