+++ src/webapp/requirements/base.txt
@@ -7,5 +7,6 @@ django-environ==0.12.0
 djangorestframework==3.16.0
 djangorestframework_gis
 gunicorn==23.0.0
-psycopg[binary]==3.2.6
+psycopg[binary]
+requests==2.32.3
 whitenoise==6.11.0
